Fail for large expiry

+++ programs/engine_testapp/mock_server.h
@@ -16,7 +16,7 @@ ServerApi* get_mock_server_api();
 
 void init_mock_server();
 
-void mock_time_travel(int by);
+void mock_time_travel(uint64_t by);
 
 void mock_set_pre_link_function(PreLinkFunction function);
